certcentral: Avoid getting stuck on CHALLENGES_PUSHED status
As seen while debugging T207737, if two orders for the same CN/SANs
are finalized in a short period of time (<2 minutes) a 400 HTTP response
is returned to certcentral, making it stuck in the CHALLENGES_PUSHED
status. This commit makes it fall back to SELF_SIGNED status in
the described scenario and restart the issuance process.

It also gets rid of the weird None return value if the challenge(s)
have not been validated yet by the ACME directory

Bug: T207737
Change-Id: I1adc5c0ab6ce0b5467c12c85c934276449680f54
(cherry picked from commit 8888aeb089c0d558268e4a055371bf36ba269fbf)