certcentral: Implement staging time
This commit adds the concept of staging time to certcentral. The
staging time it's the number of seconds that certcentral will sit on
a certificate before moving it from new_certs to live_certs. By default
is set to 3600 seconds and as Let's Encrypt backdates the certificates
1 hour, the deploy should be immediate for those certificates using
the default value.

Bug: T213737
Change-Id: I6fff2c1ab44189f7f319b91b7423030f2c182f66
(cherry picked from commit 3fa7a5d8572b6ed8cb1353310fbeead68620a91a)