acme_chief: Prevalidate CN/SNI list
This commit adds two new settings at the certificate level:
* prevalidate (True|False): When set to true the CN and the SNI list
will be prevalidated to attempt to determine if the challenges will
be able to be validated.
* skip_invalid_snis (True|False): When set to True, SNIs that fail to
prevalidate will be skipped. Otherwise certificate issuance will be
stopped at that point. Failing to prevalidate the CN will always
cause to stop the certificate issuance process.

The following validations have been included:
* All challenges: CAA record validation.
* DNS-01: NS record validation.

Bug: T220518
Change-Id: I2c965028e151df51536b6764163edd90b5ce77a7