requests: Use POST-as-GET to fetch the issued certificate
As of December 4th 2019 Let's Encrypt deprecated some GET endpoints
forcing to use POST-as-GET instead.
On acme-chief every API request is directly handled by the acme library
but it fetches the certificate using a custom method to be able to split
finalizing the order from fetching the certificate.
This fetch_certificate method needed an upgrade :)

Upstream switched the analog request from GET to POST-as-GET as part of
https://github.com/certbot/certbot/commit/0b5468e992ab57fa028ddf33ca2351cb37c8e1ee
released in version 0.29.

Bug: T244236
Change-Id: I2dcae5d51bda3d33f28a9b323d9c0fc10aa84b74