setup.py: update dependencies for bullseye
This commit is in preparation for the acme-chief hosts to bullseye. We
first start by updating the dependencies to reflect the bullseye
versions and then will later follow up with the Debian package updates.

Also update tox.ini for using Python 3.9 (default in bullseye) instead
of the current Python 3.7.

Bonus: get rid of deprecated pylint options

Bug: T321309
Change-Id: If7452cf5ab51a726bfd924cffbfa32396c574b11